requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.28.0
psycopg2-binary==2.9.9

# HTTP Client
aiohttp==3.9.0
httpx==0.25.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.9.1
isort==5.12.0
mypy==1.5.1

# Documentation
fastapi-pagination==0.12.8

# M-Pesa Integration
requests==2.31.0
pycryptodome==3.19.0

# Email
python-http-client==3.3.7
sendgrid==6.11.0

# CORS
python-multipart==0.0.6

# Async
anyio==3.7.1

# Logging
structlog==23.1.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.9.1
isort==5.12.0
mypy==1.5.1
flake8==6.1.0

# Type Hints
typing-extensions==4.8.0aiosmtplib==3.0.1
